# openapi.yaml
openapi: 3.0.3
info:
  title: StudySync API
  description: Backend API for the StudySync collaborative learning platform. Reliability score endpoints are not included.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /workspaces/{groupId}:
    get:
      tags: [Workspaces]
      summary: Workspace metadata and members
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /workspaces/{groupId}/messages:
    get:
      tags: [Workspaces]
      summary: List chat messages
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
    post:
      tags: [Workspaces]
      summary: Send message (text, attachment, or voice)
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [text]
                content:
                  type: string
          multipart/form-data:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [attachment, voice]
                file:
                  type: string
                  format: binary
                durationSec:
                  type: integer
      responses:
        '201':
          description: Created
...